test: move #12216 tests into coverage.test.ts
Not a true regression (never worked), so per convention these belong next
to the other coverage CLI tests rather than under test/regression/issue/.
